+++ src/main/java/github/hqn03/auth_service/dto/user/CreateUserRequest.java
@@ -6,5 +6,5 @@ public record CreateUserRequest(
         String username,
         String email,
         String password,
-        Set<Long> roleIds) {
+        Set<Integer> roleIds) {
 }
