+++ src/main/java/github/hqn03/auth_service/dto/user/UpdateUserRequest.java
@@ -5,5 +5,5 @@ import java.util.Set;
 public record UpdateUserRequest(
         String username,
         String email,
-        Set<Long> roleIds) {
+        Set<Integer> roleIds) {
 }
