+++ src/main/resources/db/migration/V4__create_user_roles_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE user_roles
 (
     user_id BIGINT NOT NULL,
-    role_id BIGINT NOT NULL,
+    role_id INT NOT NULL,
     PRIMARY KEY (user_id, role_id),
 
     CONSTRAINT fk_user_roles_user
