+++ src/main/resources/db/migration/V5__create_role_permissions_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE role_permissions
 (
-    role_id       BIGINT NOT NULL,
-    permission_id BIGINT NOT NULL,
+    role_id       INT NOT NULL,
+    permission_id INT NOT NULL,
     PRIMARY KEY (role_id, permission_id),
 
     CONSTRAINT fk_role_permissions_role
